feat(jwt): added jwt auth for rest endpoints, graph stubbing
