+++ graph/schemas/brigade.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'brigade',
+  description: 'Brigade schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id','auth']
+}
