+++ graph/schemas/checkins.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'checkins',
+  description: 'Checkins schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id']
+}
