+++ graph/schemas/events.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'events',
+  description: 'Events schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id']
+}
