+++ graph/schemas/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  brigade: require('./brigade.graph'),
+  projects: require('./projects.graph'),
+  users: require('./users.graph'),
+  events: require('./events.graph'),
+  posts: require('./posts.graph'),
+  checkins: require('./checkins.graph')
+}
