+++ graph/schemas/posts.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'posts',
+  description: 'Posts schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id','sync']
+}
