+++ graph/schemas/projects.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'projects',
+  description: 'Project schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id','oldId']
+}
