+++ graph/schemas/users.graph.js
@@ -1,0 +1,6 @@
+module.exports = {
+  name: 'users',
+  description: 'User schema',
+  class: 'GraphQLObjectType',
+  exclude: ['_id','tokens','jwt']
+}
