+++ helpers/bootstrapDatabase.js
@@ -23,6 +23,7 @@ module.exports = function (defaultBrigadeData, cb) {
     if (!results) {
       console.log('defaultBrigadeData', defaultBrigadeData)
       defaultBrigadeData = defaultBrigadeData || require('../seeds/Brigade')()[0]
+      console.log(defaultBrigadeData)
       defaultBrigadeData.slug = process.env.BRIGADE
       var defaultBrigade = new Brigade(defaultBrigadeData)
       defaultBrigade.save(function (err) {
