+++ models/Posts.js
@@ -8,24 +8,7 @@ var slug = require('slug')
 var getGhFile = require('github-get')
 var _ = require('lodash')
 
-var postsSchema = new mongoose.Schema({
-  slug: {type: String, default: ''}, // this is the slug
-  title: {type: String, required: true}, // Display title
-  author: {type: String, default: ''},
-  url: {type: String, default: ''}, // an external link you can use to override where to go when clicking
-  image: {type: String, default: 'http://i.imgur.com/2lHqtJ7.png'},
-  description: {type: String, default: ''},
-  content: {type: String, default: ''},
-  date: {type: Date},
-  published: {type: Boolean, default: false},
-  unix: {type: Number, default: 0},
-  sync: {
-    jekyll: {type: String, default: ''},
-    wordpress: {type: String, default: ''},
-    medium: {type: String, default: ''}
-  },
-  tags: {type: Array, default: []}
-})
+var postsSchema = require('../schemas/posts')
 
 postsSchema.statics.syncJekyll = function (blogLocation, accessToken) {
   var self = this
