+++ models/Projects/publishToGithub.js
@@ -1,0 +1,3 @@
+module.exports = function publishToGithub (cb) {
+  cb(null, 'isMatch')
+}
