+++ package.json
@@ -73,15 +73,20 @@
     "errorhandler": "^1.4.2",
     "express": "^4.13.3",
     "express-flash": "^0.0.2",
+    "express-graphql": "^0.6.1",
+    "express-jwt": "^5.1.0",
     "express-session": "^1.13.0",
     "express-validator": "^2.18.0",
     "faker": "^3.1.0",
     "gh-ls": "^1.0.0",
     "github-get": "^0.3.1",
+    "graphql": "^0.8.2",
+    "graphql-tools": "^0.8.4",
     "gulp": "^3.9.1",
     "gulp-nodemon": "^2.1.0",
     "gulp-watch": "^4.3.6",
     "is-url": "^1.2.2",
+    "jsonwebtoken": "^7.2.1",
     "link-header-parser": "0.0.2",
     "lodash": "^4.0.0",
     "lusca": "^1.3.0",
@@ -91,6 +96,7 @@
     "moment": "^2.11.1",
     "moment-timezone": "^0.5.3",
     "mongoose": "^4.3.5",
+    "mongoose-schema-to-graphql": "^2.4.5",
     "morgan": "^1.6.1",
     "node-sass-middleware": "^0.9.7",
     "node-uuid": "^1.4.7",
@@ -98,6 +104,8 @@
     "nodemailer": "^2.0.0",
     "passport": "0.3.2",
     "passport-github": "^1.0.0",
+    "passport-http-bearer": "^1.0.1",
+    "passport-jwt": "^2.2.1",
     "passport-local": "^1.0.0",
     "passport-meetup": "^1.0.0",
     "pug": "^2.0.0-beta6",
@@ -139,6 +147,7 @@
     ],
     "ignore": [
       "**/js/lib/",
+      "**/graph/",
       "migrations/"
     ]
   },
