+++ schemas/index.js
@@ -1,0 +1,8 @@
+module.exports = {
+  brigade: require('./brigade'),
+  projects: require('./projects'),
+  users: require('./users'),
+  events: require('./events'),
+  posts: require('./posts'),
+  checkins: require('./checkins')
+}
