+++ schemas/projects.js
@@ -1,4 +1,6 @@
-module.exports = {
+const mongoose = require('mongoose')
+
+module.exports = new mongoose.Schema({
   id: {type: String, default: ''}, // this is the slug - civic.sf.json + civic.dc.json
   oldId: {type: String, default: ''}, // this the old slug for updating name/id (needed in user team connections)
   brigade: {type: String, default: ''}, // this is the brigade the project currently belongs to - civic.sf.json
@@ -34,4 +36,4 @@ module.exports = {
   links: {type: Array, default: []}, // simple strings
   videos: {type: Array, default: []},
   published: {type: Boolean, default: true}
-}
+})
