+++ seeds/Brigade.js
@@ -35,7 +35,6 @@ module.exports = function () {
         type: 'permanent'
       }
     ],
-    theme: false,
     copy: {
       tagline: 'is a community searching for technological solutions to social challenges.',
       description: "We're a bunch of civic-minded technologists, designers, and topic experts using our skills to improve Example and the world."
